test(card): cover automated Gnosis flows
