+++ test/features/gnosis_card/gnosis_card_test_helpers.dart
@@ -6,38 +6,54 @@ import 'package:web_dex/features/gnosis_card/domain/gnosis_card_models.dart';
 import 'package:web_dex/features/gnosis_card/domain/gnosis_card_ports.dart';
 
 class GnosisTestSigner implements SmartAccountSigner {
+  SmartAccountOwner activeOwner = const SmartAccountOwner(
+    address: '0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+    coin: 'GNO',
+  );
   Completer<void>? personalSignatureGate;
   final Completer<void> personalSignatureStarted = Completer<void>();
   int personalSignatureCalls = 0;
   int typedDataSignatureCalls = 0;
   final List<String> registeredSafes = [];
+  final List<String> personalMessages = [];
+  final List<SmartAccountOwner> expectedPersonalOwners = [];
+  final List<SmartAccountOwner> expectedRegistrationOwners = [];
+  final List<SmartAccountOwner> expectedTypedDataOwners = [];
 
   @override
-  Future<SmartAccountOwner> owner() async => const SmartAccountOwner(
-    address: '0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-    coin: 'GNO',
-  );
+  Future<SmartAccountOwner> owner() async => activeOwner;
 
   @override
-  Future<void> registerSafe(String safeAddress) async {
+  Future<void> registerSafe(
+    String safeAddress, {
+    required SmartAccountOwner expectedOwner,
+  }) async {
+    expectedRegistrationOwners.add(expectedOwner);
     registeredSafes.add(safeAddress);
   }
 
   @override
-  Future<String> signPersonalMessage(String message) async {
+  Future<String> signPersonalMessage(
+    String message, {
+    required SmartAccountOwner expectedOwner,
+  }) async {
     personalSignatureCalls += 1;
+    personalMessages.add(message);
+    expectedPersonalOwners.add(expectedOwner);
     if (!personalSignatureStarted.isCompleted) {
       personalSignatureStarted.complete();
     }
     await personalSignatureGate?.future;
-    return 'test-eip191-signature';
+    return 'test-eip191-${message.hashCode}';
   }
 
   @override
   Future<SmartAccountSignature> signTypedData(
-    PreparedSmartAccountIntent intent,
-  ) async {
+    PreparedSmartAccountIntent intent, {
+    required SmartAccountOwner expectedOwner,
+  }) async {
     typedDataSignatureCalls += 1;
+    expectedTypedDataOwners.add(expectedOwner);
     return const SmartAccountSignature(
       signature: 'test-eip712-signature',
       typedDataHash:
@@ -68,11 +84,23 @@ class RecordingExternalFlowLauncher implements ExternalFlowLauncher {
 
 class GnosisTestPaymentGateway implements CardOrderPaymentGateway {
   int calls = 0;
+  final Map<String, CardOrderPaymentReceipt> _receipts = {};
 
   @override
-  Future<CardOrderPaymentReceipt> pay(CardOrderPaymentQuote quote) async {
+  Future<CardOrderPaymentReceipt?> findPayment({
+    required CardOrderPaymentQuote quote,
+    required String idempotencyKey,
+  }) async => _receipts[idempotencyKey];
+
+  @override
+  Future<CardOrderPaymentReceipt> pay(
+    CardOrderPaymentQuote quote, {
+    required String idempotencyKey,
+  }) async {
+    final existing = _receipts[idempotencyKey];
+    if (existing != null) return existing;
     calls += 1;
-    return CardOrderPaymentReceipt(
+    final receipt = CardOrderPaymentReceipt(
       orderId: quote.orderId,
       transactionHash: '0xtest-synthetic-receipt',
       amountMinor: quote.amountMinor,
@@ -80,6 +108,8 @@ class GnosisTestPaymentGateway implements CardOrderPaymentGateway {
       paidAt: DateTime.utc(2026, 7, 14, 12),
       isSimulated: true,
     );
+    _receipts[idempotencyKey] = receipt;
+    return receipt;
   }
 }
